requirements.txt
langchain
python-dotenv
llama-cpp-python
langchain_community
